# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do controle de estoque para pizzarias
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputUnit } from './productInputUnit';

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @minimum 1 */
  categoryId: number;
  /**
     * @minimum 1
     * @nullable
     */
  supplierId?: number | null;
  ca: string;
  manufacturer: string;
  size?: string | null;
  unit: ProductInputUnit;
  /** @minimum 0 */
  currentStock: number;
  /** @minimum 0 */
  minimumStock: number;
  /** @minimum 0 */
  maximumStock: number;
  /** @minimum 0 */
  unitCost: number;
}